New calculator functions

+++ calculator.go
@@ -1,3 +1,3 @@
 package CalcLib
 
-type Calculator interface{ Calculate(a, b int) int }
+type Calculator interface{ Calculate(a, b float64) float64 }
